+++ security/forms.py
@@ -2,6 +2,7 @@ from django import forms
 from django.utils.translation import gettext_lazy as _
 from django.contrib.auth.password_validation import validate_password
 from django.forms import ValidationError
+from django.contrib.auth.forms import PasswordResetForm
 
 from .models import User
 from .constants import SIGN_UP_REQUEST_ID
@@ -151,3 +152,16 @@ class SignInForm(forms.Form):
             }
         )
     )
+
+
+class ResetPasswordForm(PasswordResetForm):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        self.fields["email"].widget.attrs.update(
+            {
+                "class": "block w-full h-10 rounded-md border-0 py-1.5 px-3 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-1 focus:ring-inset focus:ring-gray-300 sm:text-sm sm:leading-6 focus-visible:outline-transparent",
+                "placeholder": "Enter your email",
+                "required": True,
+            }
+        )
