+++ security/templates/security/reset_password.html
@@ -36,7 +36,7 @@
                                 In order to reset your password, you need to enter your email.
                             </div>
 
-                            {% include 'talent/reset_password_form.html' %}
+                            {% include 'security/reset_password_form.html' %}
 
                             <p class="mt-6 text-left text-sm text-gray-500">
                                 Don’t have an account yet?
