+++ security/templates/security/reset_password_form.html
