+++ security/templates/security/sign_up_step_2.html
@@ -32,7 +32,7 @@
         <div>
             <label for="{{ wizard.form.password.id_for_label }}"
                 class="block text-sm font-medium leading-6 text-gray-900">{{ wizard.form.password.label }}</label>
-            <!-- Note that the below password field is assigned to an id in talent/forms.py
+            <!-- Note that the below password field is assigned to an id in security/forms.py
                 which will be used in Javascript to check the password strength-->
             <div class="mt-2">
                 {{ wizard.form.password }}
