--- talent/forms.py
@@ -1,14 +1,0 @@
-from django.contrib.auth.forms import PasswordResetForm
-
-
-class ResetPasswordForm(PasswordResetForm):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        self.fields["email"].widget.attrs.update(
-            {
-                "class": "block w-full h-10 rounded-md border-0 py-1.5 px-3 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-1 focus:ring-inset focus:ring-gray-300 sm:text-sm sm:leading-6 focus-visible:outline-transparent",
-                "placeholder": "Enter your email",
-                "required": True,
-            }
-        )
